+++ packages/ui/src/constants/constants.ts
@@ -1,4 +1,125 @@
+import { Home as HomeIcon, Plus, Settings, TriangleAlert } from "lucide-react";
+
+export const USER_ROLE: "user" | "admin" = "user";
+
+export const USER_PAGE_LINKS = [
+  {
+    title: "Dashboard",
+    url: "/",
+    icon: HomeIcon,
+  },
+  {
+    title: "New Sentiment",
+    url: "/new-sentiment",
+    icon: Plus,
+  },
+];
+
+export const ADMIN_PAGE_LINKS = [
+  {
+    title: "Dashboard",
+    url: "/",
+    icon: HomeIcon,
+  },
+  {
+    title: "Alerts",
+    url: "/alerts",
+    icon: TriangleAlert,
+  },
+  {
+    title: "Settings",
+    url: "/settings",
+    icon: Settings,
+  },
+];
+
 // Add manual feedback page
+export const TEST_PREVIEW_FEEDBACKS = [
+  {
+    id: "f4be7d2d-8619-44d8-9132-d19cf59e1bfa",
+    feedback: "This app saved me hours every week. Super intuitive and fast!",
+  },
+  {
+    id: "b64ac73e-dcfd-4ea1-b168-6f490b88d72b",
+    feedback: "I love how seamless everything feels. Great UX!",
+  },
+  {
+    id: "f91d87fc-bd00-4a3d-b4ec-b6107a7ec94d",
+    feedback:
+      "The onboarding was smooth and easy. Highly recommend to beginners.",
+  },
+  {
+    id: "e26796f2-d5f3-4d3b-a755-d63b004c2ff4",
+    feedback:
+      "Customer support was incredibly responsive. Fixed my issue within minutes.",
+  },
+  {
+    id: "2aa3f1dc-13e8-44ee-9f08-0f7b979e7081",
+    feedback: "Reliable, clean, and efficient. What more could you ask for?",
+  },
+  {
+    id: "416740e6-5f0c-4c20-8d42-6991032b5244",
+    feedback: "We switched from another tool and haven’t looked back.",
+  },
+  {
+    id: "f1961c62-f0bc-41a2-b1f2-2a4a59ecdd32",
+    feedback: "Beautiful design and very well thought out features.",
+  },
+  {
+    id: "a0d3b745-23d0-4171-9c95-1a602d2b3c65",
+    feedback: "A bit of a learning curve, but once you get it, it’s powerful.",
+  },
+  {
+    id: "7d43e3dc-087b-4d44-a305-118b365d46f4",
+    feedback: "Works great on both desktop and mobile. Love the consistency!",
+  },
+  {
+    id: "ed15e512-7b45-4cfd-8bc5-3a987bb209e1",
+    feedback:
+      "Had one bug but it was fixed quickly after I reported it. Great devs!",
+  },
+  {
+    id: "de5f285b-002e-4493-8843-f1d222dc4de2",
+    feedback: "Honestly exceeded my expectations. Super smooth experience.",
+  },
+  {
+    id: "7e264b1d-5b89-4fbe-975a-2d94a999c859",
+    feedback: "Perfect tool for remote teams. We’re using it every day.",
+  },
+  {
+    id: "e9c42dfb-5823-404c-83e6-e8121cf2836e",
+    feedback: "Speed and simplicity — this app nails both.",
+  },
+  {
+    id: "37a90bd0-2d8a-4c2f-bc04-f82c2732a71c",
+    feedback: "I rarely leave reviews, but this one deserves it.",
+  },
+  {
+    id: "a3d00599-13a9-452e-9d33-f68d3123e6db",
+    feedback: "All the features I need, none of the clutter. Impressive!",
+  },
+  {
+    id: "5b2d5e39-56b9-4c8f-a5c9-c3a202d9832f",
+    feedback: "Really improved our workflow. Thanks for building this!",
+  },
+  {
+    id: "347ef2d0-5ef5-4bb2-a4a2-3c6d14a64020",
+    feedback: "Clean UI, blazing fast, and easy to navigate.",
+  },
+  {
+    id: "f2b7edcb-b15f-47eb-82e7-ecbe1a300f94",
+    feedback: "I use this daily and it’s never let me down.",
+  },
+  {
+    id: "12d0dc3d-1d10-4d14-85a2-2cd18b7d68f0",
+    feedback: "Just what I needed. Saves me so much time.",
+  },
+  {
+    id: "da786fa2-96db-434b-b362-c442c55f07c8",
+    feedback: "They keep updating it with useful features. Love the roadmap.",
+  },
+];
+
 export const FEEDBACK_MIN_LENGTH = 10;
 export const FEEDBACK_MAX_LENGTH = 3000;
 export const FEEDBACK_MAX_ITEMS = 100;
